requirements.txt
# Core
vapi-server-sdk>=1.10.0
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Dev
pytest>=8.3.0
pytest-asyncio>=0.23.0
httpx  # also used as test client for FastAPI
